requirements.txt
numpy>=1.21.0
tensorflow-gpu>=2.7.0
pretty_midi>=0.2.9
music21>=7.3.3
PyQt6>=6.2.0
midiutil>=1.2.1
dataclasses-json>=0.5.7  # For JSON serialization of dataclasses
python-dateutil>=2.8.2  # For datetime handling
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
pyinstaller>=6.3.0
librosa>=0.10.1  # For advanced music analysis
transformers>=4.30.0  # For transformer models
pytorch-lightning>=2.0.0  # For better model training
torch>=2.0.0  # PyTorch for deep learning
tensorboard>=2.14.0  # For training visualization
mir_eval>=0.7  # For music information retrieval tasks